==> 03c - Engineered Data Quality Report.ipynb <==
# Engineered Data Quality Report

In this notebook we are going to create a data quality report piece-by-piece and then put it all together at the end.

# Import the libraries we need
import pandas as pd
import pyreadr

# Import the data from the source CSV file the 'Create a Pandas DataFrame From a CSV File' recipe and take a peek
credit_card_default_raw = pd.read_csv('./Data/credit_card_default.csv')
result = pyreadr.read_r('./Data/credit_card_default_eng.RData') 
credit_card_default_eng = result[None]
credit_card_default = credit_card_default_eng.copy()
df_flags = credit_card_default_raw['data.group']
credit_card_default = credit_card_default.join(df_flags)

credit_card_default.head()

## Available Columns

# Create a DataFrame of the columns in the credit_card_default dataframe
columns = pd.DataFrame(list(credit_card_default.columns.values))
columns

## Data Types

# Create a DataFrame of the data type of each column
data_types = pd.DataFrame(credit_card_default.dtypes,
                          columns=['Data Type'])
data_types

## Count of Missing Values in Each Column

# Create a DataFrame with the count of missing values in each column
missing_data_counts = pd.DataFrame(credit_card_default.isnull().sum(),
                                   columns=['Missing Values'])
missing_data_counts

## Count of Present Values in Each Column

# Create a DataFrame with the count of present values in each column
present_data_counts = pd.DataFrame(credit_card_default.count(),
                                   columns=['Present Values'])
present_data_counts

## Number of Unique Values Per-Column

# Create DataFrame with the count of unique values in each column
unique_value_counts = pd.DataFrame(columns=['Unique Values'])
for v in list(credit_card_default.columns.values):
    unique_value_counts.loc[v] = [credit_card_default[v].nunique()]
unique_value_counts

## The Minimum Value In Each Column

credit_card_default.age_bins = credit_card_default.age_bins.values.as_ordered() # need to convert

# Create a DataFrame with the minimum value in each column
minimum_values = pd.DataFrame(columns=['Minimum Value'])
for v in list(credit_card_default.columns.values):
    minimum_values.loc[v] = [credit_card_default[v].min()]
minimum_values

## The Maximum Value In Each Column

# Create a DataFrame with the minimum value in each column
maximum_values = pd.DataFrame(columns=['Maximum Value'])
for v in list(credit_card_default.columns.values):
    maximum_values.loc[v] = [credit_card_default[v].max()]
maximum_values

# Bring It All Together

# Merge all the DataFrames together by the index
data_quality_report = data_types.join(present_data_counts).join(missing_data_counts).join(unique_value_counts).join(minimum_values).join(maximum_values)

# Print out a nice report
print("\nData Quality Report")
print("Total records: {}".format(len(credit_card_default.index)))
data_quality_report
